Move 'FeedRefreshViewController' to a separate file
