+++ MVVM Module/Source/Feed UI/Controllers/FeedViewController.swift
@@ -4,34 +4,6 @@
 
 import UIKit
 
-final class FeedRefreshViewController: NSObject {
-    @IBOutlet var view: UIRefreshControl?
-    @IBOutlet var errorView: ErrorView?
-    
-    var viewModel: FeedViewModel? {
-        didSet { bind() }
-    }
-
-    @IBAction func refresh() {
-        viewModel?.loadFeed()
-    }
-    
-    private func bind() {
-        viewModel?.onLoadingStateChange = { [weak self] isLoading in
-            if isLoading {
-                self?.errorView?.hideMessage()
-                self?.view?.beginRefreshing()
-            } else {
-                self?.view?.endRefreshing()
-            }
-        }
-        
-        viewModel?.onFeedFailed = { [weak self] message in
-            self?.errorView?.show(message: message)
-        }
-    }
-}
-
 public final class FeedViewController: UITableViewController, UITableViewDataSourcePrefetching {
     @IBOutlet var refreshViewController: FeedRefreshViewController?
